use model and View
